% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap.R
\name{is_leap}
\alias{is_leap}
\title{Check if a year is a leap year.}
\usage{
is_leap(year)
}
\arguments{
\item{year}{Year to input}
}
\value{
Boolean
}
\description{
Check if a year is a leap year.
}
\examples{
is_leap(2000)
#> TRUE
is_leap(1900)
#> FALSE

}
