% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conv.R
\name{temp_conv}
\alias{temp_conv}
\title{Convert temperature values}
\usage{
temp_conv(temp, start_scale, end_scale)
}
\arguments{
\item{temp}{Temperature in given start_scale}

\item{start_scale}{Temperature scale for start; "F" for Farenheit, "C" for Celcius, "K" for Kelvin}

\item{end_scale}{Temperature scale for end; "F" for Farenheit, "C" for Celcius, "K" for Kelvin}
}
\value{
Temperature in given end_scale
}
\description{
Convert temperature values
}
\examples{
temp_conv(41, "F", "C")
#> 5
temp_conv(5, "C", "K")
#> 278.15
temp_conv(278.15, "K", "F")
#> 41

}
